FUSETOOLS2-574 - General cleanup of unused imports, access modifiers and others. Moved FileType to its own class
