+++ src/main/java/com/github/cameltooling/lsp/internal/parser/CamelKModelineInsertionParser.java
@@ -1,14 +1,9 @@
 package com.github.cameltooling.lsp.internal.parser;
 
+import com.github.cameltooling.lsp.internal.completion.modeline.CamelKModelineFileType;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.TextDocumentItem;
 
-import java.io.File;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -30,7 +25,7 @@ public class CamelKModelineInsertionParser {
     }
 
     private boolean isCamelKFile() {
-        return FileType.getFileTypeCorrespondingToUri(document.getUri()).isPresent();
+        return CamelKModelineFileType.getFileTypeCorrespondingToUri(document.getUri()).isPresent();
     }
 
     private boolean lineIsEmpty(int line) {
@@ -46,63 +41,7 @@ public class CamelKModelineInsertionParser {
                 .map(currLine -> new ParserFileHelperUtil().getLine(document,currLine))
                 .collect(Collectors.joining("\n"));
 
-        return FileType.getFileTypeCorrespondingToUri(document.getUri()).orElseThrow()
+        return CamelKModelineFileType.getFileTypeCorrespondingToUri(document.getUri()).orElseThrow()
                 .checkTextIsCommentsDelegate.apply(textBeforeLine);
-    }
-
-    private enum FileType {
-        XML(List.of(".camelk.xml"), "<!-- camel-k:", FileType::textIsFullyCommentedXML),
-        Java(List.of(".java"), "// camel-k:", FileType::textIsFullyCommentedJava),
-        YAML(List.of(".camelk.yaml", ".camelk.yml"), "# camel-k", FileType::textIsFullyCommentedYAML);
-
-        public final List<String> extensions;
-        public final String modelineLabel;
-        public final Function<String, Boolean> checkTextIsCommentsDelegate;
-
-        private FileType(List<String> extensions, String modelineLabel, Function<String, Boolean> checkTextIsCommentsDelegate) {
-            this.extensions = extensions;
-            this.modelineLabel = modelineLabel;
-            this.checkTextIsCommentsDelegate = checkTextIsCommentsDelegate;
-        }
-
-        private static Optional<FileType> getFileTypeCorrespondingToUri(String uri) {
-            return Arrays.asList(FileType.values()).stream()
-                    .filter(type ->
-                        type.extensions.stream().anyMatch(uri::endsWith)
-                    )
-                    .findFirst();
-        }
-
-        private static boolean textIsFullyCommentedXML(String text){
-            //Remove all segments between <!-- and -->. Check if it's empty.
-            Pattern commentRegex = Pattern.compile("<!--.*-->");
-
-            return textIsFullOfRegex(text, commentRegex);
-        }
-
-        private static boolean textIsFullyCommentedYAML(String text){
-            //Remove all segments between # and \n
-            Pattern commentRegex = Pattern.compile("#.*\\n");
-
-            return textIsFullOfRegex(text, commentRegex);
-        }
-
-        private static boolean textIsFullyCommentedJava(String text){
-            //Line Comments: Remove from // to \n
-            //Block Comments: Remove from /* to */. Newlines have to be explicitly added
-            Pattern lineComment = Pattern.compile("\\/\\/.*\\n");
-            Pattern blockComment = Pattern.compile("\\/\\*(.|\\n)*\\*\\/");
-            Pattern commentRegex = Pattern.compile(String.format("(%s|%s)",lineComment.pattern(), blockComment.pattern()));
-
-            return textIsFullOfRegex(text, commentRegex);
-        }
-
-        private static boolean textIsFullOfRegex(String text, Pattern regex) {
-            //Add an extra carriage return at the end for correct matching with line comments
-            String textWithExtraCarriageReturn = text + "\n";
-            String textWithoutComments = textWithExtraCarriageReturn.replaceAll(regex.pattern(), "");
-
-            return textWithoutComments.isBlank();
-        }
     }
 }
